requirements.txt
boto3==1.35.55
confluent-kafka>=2.5.0
fastapi>=0.111.1
# faker >= 33.1.0
jsonref>=1.1.0
numpy>=1.26.4
openai>=1.47.1
pinecone-client==4.1.2
psycopg2-binary>=2.9.10
pydantic>=2.8.2
pymilvus>=2.4.6
pymupdf>=1.24.9
python-multipart>=0.0.9
pyyaml>=6.0.1
redis>=5.0.8
requests>=2.32.3
typing==3.7.4.3
#utils==1.0.2 # TODO: Conflict with utils from coreagents
uvicorn>=0.30.3
streamlit
PyPDF2
python-docx
python-pptx
fpdf
pandas
openpyxl
xhtml2pdf
# testing
pytest>=8.3.2
pylint>=3.2.6
